requirements.txt
discord.py[voice]
humanize
jikanpy
jishaku
psutil
async_cleverbot
git+https://github.com/Rapptz/discord-ext-menus